Reorganize
Add coverage report to clean script
