+++ bin/clean
@@ -1,2 +1,3 @@
 #!/usr/bin/env bash
-rm -rf build dist mpr.egg-info MANIFEST .coverage coverage ./**/__pycache__
+rm -rf build coverage dist mpr.egg-info MANIFEST .coverage coverage ./**/__pycache__
+
