+++ mpr/cash_index.py
@@ -1,23 +1,23 @@
 from argparse import ArgumentParser
 from asyncio import run
-
 from datetime import date
 from datetime import timedelta
 from functools import singledispatch
+
 from pandas import DataFrame
 
-from .slaughter import get_slaughter
+from . import lm_hg201
 from .slaughter.cash_index import cash_index_report
 
 
 @singledispatch
 async def get(start: date, end=date.today()) -> DataFrame:
-    slaughter = await get_slaughter(start - timedelta(days=5), end)
+    slaughter = await lm_hg201.fetch(start - timedelta(days=5), end)
     return cash_index_report(slaughter)[start:]
 
 
 @get.register(int)
-async def get_recent_cash_prices(n: int) -> DataFrame:
+async def get_recent(n: int) -> DataFrame:
     today = date.today()
     cash_prices = await get(today - timedelta(n + 10), today)
     return cash_prices.tail(n)
