+++ mpr/cuts/model.py
@@ -9,11 +9,18 @@ from numpy import rec
 from numpy import uint32
 from numpy import uint8
 
-from . import CutType
+from .cut_type import CutType
+from .cut_type import cut_types
 from ..data import Date
+from ..data import Record
 from ..data import date64
 from ..data import date_ordinal
 from ..data import unicode
+from ..data import opt_float
+from ..data import opt_int
+from ..data import parse_date
+
+date_format = "%m/%d/%Y"
 
 
 class Cut(NamedTuple):
@@ -47,3 +54,22 @@ def to_array(records: Iterator[Cut]) -> recarray:
         ('low_price', float32),
         ('high_price', float32)
     ]))
+
+
+def parse_record(record: Record) -> Cut:
+    report = record['slug'].lower()
+    report_date = parse_date(record['report_date'], date_format)
+    section = record['label']
+    cut_type = cut_types[section].value
+    description = record['Item_Description']
+
+    return Cut(
+        report=report,
+        date=report_date,
+        report_date=report_date,
+        type=cut_type,
+        description=description,
+        weight=opt_int(record, 'total_pounds'),
+        avg_price=opt_float(record, 'weighted_average'),
+        low_price=opt_float(record, 'price_range_low'),
+        high_price=opt_float(record, 'price_range_high'))
