+++ mpr/purchase_index.py
@@ -1,19 +1,17 @@
+from argparse import ArgumentParser
+from asyncio import run
 from datetime import date
 from datetime import timedelta
 from functools import singledispatch
 from pandas import DataFrame
 
-from mpr import purchase
-from .purchase.report import purchase_report
-
-
-from argparse import ArgumentParser
-from asyncio import run
+from . import lm_hg200
+from .purchase.purchase_index import purchase_report
 
 
 @singledispatch
 async def get(start: date, end=date.today()) -> DataFrame:
-    records = await purchase.prior_day(start - timedelta(10), end)
+    records = await lm_hg200.fetch(start - timedelta(10), end)
     return purchase_report(records)[start:]
