+++ mpr/slaughter/purchase_type.py
@@ -1,0 +1,32 @@
+from typing import Dict
+
+from ..purchase_type import Arrangement
+from ..purchase_type import Basis
+from ..purchase_type import PurchaseType
+from ..purchase_type import Seller
+
+purchase_types: Dict[str, PurchaseType] = {
+    'Prod. Sold Negotiated':
+        (Seller.PRODUCER, Arrangement.NEGOTIATED, Basis.ALL),
+
+    'Prod. Sold Swine or Pork Market Formula':
+        (Seller.PRODUCER, Arrangement.MARKET_FORMULA, Basis.ALL),
+
+    'Prod. Sold Other Market Formula':
+        (Seller.PRODUCER, Arrangement.OTHER_MARKET_FORMULA, Basis.ALL),
+
+    'Prod. Sold Negotiated Formula':
+        (Seller.PRODUCER, Arrangement.NEGOTIATED_FORMULA, Basis.ALL),
+
+    'Prod. Sold Other Purchase Arrangement':
+        (Seller.PRODUCER, Arrangement.OTHER_PURCHASE, Basis.ALL),
+
+    'Prod. Sold (All Purchase Types)':
+        (Seller.PRODUCER, Arrangement.ALL, Basis.ALL),
+
+    'Pack. Sold (All Purchase Types)':
+        (Seller.PACKER, Arrangement.ALL, Basis.ALL),
+
+    'Packer Owned':
+        (Seller.PACKER, Arrangement.PACKER_OWNED, Basis.ALL)
+}
