+++ test/cutout/test_cutout_report.py
@@ -3,8 +3,8 @@ from itertools import starmap
 
 from numpy import allclose
 
-from mpr.cutout.api import parse_record
 from mpr.cutout.cutout_index import cutout_report
+from mpr.cutout.model import parse_record
 
 with open('test/resources/cutout.json') as resource:
     reports = json.load(resource)
