+++ test/cuts/test_model.py
@@ -1,4 +1,4 @@
-from mpr.cuts.api import parse_record
+from mpr.cuts.model import parse_record
 
 
 def test_objects_are_the_same():
