+++ test/data/test_repository.py
@@ -4,11 +4,11 @@ from pathlib import Path
 from pytest import fixture
 from pytest import mark
 
+from mpr import lm_pk602
+from mpr.cutout.report import CutoutReport
+from mpr.data.repository import Repository
 from mpr.data.repository import merge_reports
 from mpr.data.repository import slice_dates
-from mpr.report import CutoutReport
-from mpr.data.repository import Repository
-from mpr.report import lm_pk602
 from test.server import server
