+++ test/purchase/test_model.py
@@ -1,4 +1,4 @@
-from mpr.purchase.api import parse_record
+from mpr.purchase.model import parse_record
 
 
 def test_objects_are_the_same():
