+++ test/purchase/test_parse_record.py
@@ -3,7 +3,7 @@ from datetime import date
 from numpy import isnan
 from numpy import isclose
 
-from mpr.purchase.api import parse_record
+from mpr.purchase.model import parse_record
 from mpr.purchase.model import to_array
 from mpr.purchase_type import Seller, Arrangement, Basis
