+++ test/server.py
@@ -7,10 +7,10 @@ from aiohttp.web import FileResponse
 from aiohttp.web import get
 from aiohttp.web_app import Application
 
-from mpr.report import Report
-from mpr.report import lm_hg200
-from mpr.report import lm_hg201
-from mpr.report import lm_pk602
+from mpr import lm_hg200
+from mpr import lm_hg201
+from mpr import lm_pk602
+from mpr.data.report import Report
 
 
 def route(report: Report):
